+++ core/ViewDataTable/Config.php
@@ -463,6 +463,16 @@ class Config
     public $export_limit = '';
 
     /**
+     * Message to show if not data is available for the report
+     * Defaults to `CoreHome_ThereIsNoDataForThisReport` if not set
+     *
+     * Attention: Message will be printed raw. Don't forget to escape where needed!
+     *
+     * @var string
+     */
+    public $no_data_message = '';
+
+    /**
      * @ignore
      */
     public $report_id = '';