+++ plugins/UsersManager/angularjs/manage-user-access/manage-user-access.controller.js
@@ -43,6 +43,7 @@
         };
 
         this.setAccess = function (login, access) {
+          login = String(login);
           login=piwik.helper.escape(piwik.helper.htmlEntities(login));
             if ( $('[data-login="' + login + '"]').find("#"+access).has('.accessGranted').length ){
                 return;