+++ tests/PHPUnit/Framework/TestingEnvironmentVariables.php
@@ -105,6 +105,7 @@ class TestingEnvironmentVariables
         $disabledPlugins = $pluginList->getCorePluginsDisabledByDefault();
         $disabledPlugins[] = 'LoginHttpAuth';
         $disabledPlugins[] = 'LoginLdap';
+        $disabledPlugins[] = 'MarketingCampaignsReporting';
         $disabledPlugins[] = 'ExampleVisualization';
 
         $disabledPlugins = array_diff($disabledPlugins, array(